Introduce options formatting and pipe lists
The "options" argument will start accepting a range of values, from
strings to the previously accepted objects.

When non-objects are passed, they will be treated as the "pipes"
property of PipeOptions.

When single elements are included as pipes, they will be converted to
arrays, and when strings are given, the pipe list will come into play.

Pipe lists are predefined lists of pipes that can be sent to any of the
Plumbio's pipes. When a string is given instead of a pipe, it will be
used as a key in the search for the real pipe in the pipe list.

Last but not least, the "pipe" property in PipeOptions will start being
converted/merged into "pipes", becoming an alias for it not only
externally but internally.