Rework "then" to use getSubResult
Also use result.status as is when checking it, without using true in its
place when the property doesn't exist.

This change is inevitable because, when using getSubResult, the result will
never lack a "status" property to give true a chance to be used instead.

Since an absent/undefined status is cast to false by getSubResult, this
creates a breaking change for when status is not provided by one of the
pipes in "then".