feat(containers): accept a value in setTime, parsing it if it's an object
In this commit, setTime can optionally accept a value representing a datetime, which will make it
create a Unix timestamp based on the datetime received rather than in the current datetime. If such
value is an object, it will be parsed to create a string the JavaScript's Date() constructor can
understand.