refactor: the pipe store now lives solely in pipe options
In this commit, the practice of storing the pipe store in the stream is ceased, given that the new
default options feature make it possible to pass the store to all pipes. As a result, the
getFormattedStream helper is removed, since is exists mostly to set the pipe store in the stream and
to manipulate the stream based on the options by using a stream resolver, something that can be done
inside the pipes themselves. Finally, the new standard property name for the pipe store becomes
"pipeStore".

BREAKING CHANGE: pipe stores can no longer be set in the stream and the getFormattedStream helper is
removed.