fix: use more explicit parameters to declare pipes that are not dynamic
In this commit, pipes that are not dynamic are declared with the explicit parameters required by
their type. This fixes a bug where a Pipe<WithOptionsAndStream>, when called with only the options
argument provided, would treat it as a stream, instead of treating it as options and providing an
empty stream object.